id 자료형 수정 & Fetch type 수정.

+++ src/main/kotlin/sogong/sogongSpring/dto/HashtagDbDto.kt
@@ -3,6 +3,6 @@ package sogong.sogongSpring.dto
 import java.io.Serializable
 
 data class HashtagDbDto(
-    val hashid: Int? = null,
+    val hashid: Long? = null,
     val hashname: String? = null
 ) : Serializable
