+++ src/main/kotlin/sogong/sogongSpring/repository/BasicLoginRepository.kt
@@ -3,5 +3,5 @@ package sogong.sogongSpring.repository
 import org.springframework.data.jpa.repository.JpaRepository
 import sogong.sogongSpring.entity.BasicLoginEntity
 
-interface BasicLoginRepository : JpaRepository<BasicLoginEntity, Int> {
+interface BasicLoginRepository : JpaRepository<BasicLoginEntity, Long> {
 }
