+++ src/main/kotlin/sogong/sogongSpring/repository/EntirePostRepository.kt
@@ -3,5 +3,5 @@ package sogong.sogongSpring.repository
 import org.springframework.data.jpa.repository.JpaRepository
 import sogong.sogongSpring.entity.EntirePostEntity
 
-interface EntirePostRepository : JpaRepository<EntirePostEntity, Int> {
+interface EntirePostRepository : JpaRepository<EntirePostEntity, Long> {
 }
